restore AutoBePlaygroundReplayIndexMovie
